- Fixed signed compiler warning.

+++ src/client/gui/pawsactivemagicwindow.cpp
@@ -184,7 +184,7 @@ void pawsActiveMagicWindow::HandleMessage(MsgEntry* me)
     }
     else
     {
-        for( int i=0; i<numSpells; i++ )
+        for( size_t i=0; i<numSpells; i++ )
         {
             if(incoming.duration[i]==0 && showEffects==false)
             {
